feat: faculty page frontend
